backend/pyproject.toml
[project]
name = "nodal"
version = "0.1.0"
description = ""
authors = [
    {name = "Romanos Chiliarchopoulos",email = "romanoshiliarhopoulos@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "litellm (>=1.82.4,<2.0.0)",
    "sse-starlette (>=3.3.3,<4.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "firebase-admin (>=7.3.0,<8.0.0)",
    "google-cloud-firestore (>=2.25.0,<3.0.0)",
    "google-cloud-secret-manager (>=2.26.0,<3.0.0)",
    "cryptography (>=46.0.5,<47.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

